feat: Vehicle: add brand, regTag & location
Backend:

+ Add fields:
  + Brand
  + RegTag
  + vehicleType.batch
  + Location (relation)
- Remove unused fields

* Restructure vehicle subclasses:
  * Remove scooter type class
  + Add batch type class

* Restructure ticket subclasses
  * Move isOpen/endDateto repair and rental

Frontend:

+ Add columns, filters and forms for brand, regTag & location
* Update types, enums and validation to match all vehicle sub types
* Handle non 400 errors as errors in api layer

:)

+++ main/src/main/java/se/hjulverkstan/main/service/LocationServiceImpl.java
@@ -3,12 +3,15 @@ package se.hjulverkstan.main.service;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import se.hjulverkstan.Exceptions.CouldNotDeleteException;
 import se.hjulverkstan.Exceptions.ElementNotFoundException;
 import se.hjulverkstan.main.dto.LocationDto;
 import se.hjulverkstan.main.dto.NewLocationDto;
 import se.hjulverkstan.main.dto.responses.GetAllLocationDto;
 import se.hjulverkstan.main.model.Location;
+import se.hjulverkstan.main.model.Vehicle;
 import se.hjulverkstan.main.repository.LocationRepository;
+import se.hjulverkstan.main.repository.VehicleRepository;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,11 +20,14 @@ import java.util.List;
 @Transactional
 public class LocationServiceImpl implements LocationService {
     private final LocationRepository locationRepository;
-    public static final String ELEMENT_NAME = "Location";
+    private final VehicleRepository vehicleRepository;
+    public static final String ELEMENT_LOCATION = "Location";
+    public static final String ELEMENT_VEHICLE = "Vehicle";
 
     @Autowired
-    public LocationServiceImpl(LocationRepository locationRepository) {
+    public LocationServiceImpl(LocationRepository locationRepository, VehicleRepository vehicleRepository) {
         this.locationRepository = locationRepository;
+        this.vehicleRepository = vehicleRepository;
     }
 
     @Override
@@ -39,7 +45,7 @@ public class LocationServiceImpl implements LocationService {
     @Override
     public LocationDto getLocationById(Long id) {
         Location location = locationRepository.findById(id)
-                .orElseThrow(() -> new ElementNotFoundException(ELEMENT_NAME));
+                .orElseThrow(() -> new ElementNotFoundException(ELEMENT_LOCATION));
 
         return new LocationDto(location);
     }
@@ -47,7 +53,11 @@ public class LocationServiceImpl implements LocationService {
     @Override
     public LocationDto deleteLocation(Long id) {
         Location location = locationRepository.findById(id)
-                .orElseThrow(() -> new ElementNotFoundException(ELEMENT_NAME));
+                .orElseThrow(() -> new ElementNotFoundException(ELEMENT_LOCATION));
+
+        if (location.getVehicles() != null && !location.getVehicles().isEmpty()) {
+            throw new CouldNotDeleteException(ELEMENT_LOCATION);
+        }
 
         locationRepository.delete(location);
         return new LocationDto(location);
@@ -56,13 +66,21 @@ public class LocationServiceImpl implements LocationService {
     @Override
     public LocationDto editLocation(Long id, LocationDto location) {
         Location selectedLocation = locationRepository.findById(id)
-                .orElseThrow(() -> new ElementNotFoundException(ELEMENT_NAME));
+                .orElseThrow(() -> new ElementNotFoundException(ELEMENT_LOCATION));
 
         selectedLocation.setAddress(location.getAddress());
         selectedLocation.setName(location.getName());
         selectedLocation.setLocationType(location.getLocationType());
         selectedLocation.setComment(location.getComment());
 
+        List<Vehicle> vehicles = vehicleRepository.findAllById(location.getVehicleIds());
+
+        vehicles.forEach(vehicle -> {
+            vehicle.setLocation(selectedLocation);
+            vehicleRepository.save(vehicle);
+        });
+
+        selectedLocation.setVehicles(vehicles);
         locationRepository.save(selectedLocation);
         return new LocationDto(selectedLocation);
     }
@@ -74,7 +92,7 @@ public class LocationServiceImpl implements LocationService {
         location.setName(newLocation.getName());
         location.setLocationType(newLocation.getLocationType());
         location.setComment(newLocation.getComment());
-
+        location.setVehicles(new ArrayList<>());
         locationRepository.save(location);
         return new LocationDto(location);
     }
